lab6 changes

+++ requests/delete_image_request.js
@@ -1,13 +1,26 @@
 import express from 'express';
 import { unlink } from 'fs';
 import * as dbPictures from '../db/db_pictures.js';
+import * as dbListings from '../db/db_listings.js';
 
 const router = express.Router();
 
 // remove an image from the database and file system that matches the descriptions
 router.post('/delete_image', express.json(), async (req, res) => {
+  // if no one is logged in, redirect to front page
+  if (!req.session.sessionUser) {
+    res.status(401).redirect('/');
+    return;
+  }
   const { pictureName } = req.body;
   const { listingID } = req.body;
+  const [listings] = await dbListings.getListingByID(listingID);
+  const listing = listings[0];
+  // if logged in user is not the owner of the listing, then don't delete image
+  if (listing.Username !== req.session.sessionUser.Username) {
+    res.status(401).send({ message: "This listing doesn't belong to you" });
+    return;
+  }
   const [dbResponse] = await dbPictures.deletePicture(listingID, pictureName);
   if (dbResponse.affectedRows) {
     unlink(`./public/Pictures/${pictureName}`, () => {
